Implemented Chloe's Start Menu.
