+++ .gitignore
@@ -13,3 +13,5 @@ Cannon Prototype/Library/metadata/97/97decbdab0634cdd991f8d23ddf0dead
 Cannon Prototype/Library/metadata/00/00000000000000002000000000000000
 Cannon Prototype/Library/metadata/00/00000000000000002000000000000000
 Cannon Prototype/Library/metadata/00/00000000000000003000000000000000
+Cannon Prototype/Temp/
+Cannon Prototype/Library/metadata/
